change m to cm

+++ raspberryPI3/ros2_ws/src/car_control/src/car_control_node.cpp
@@ -158,8 +158,8 @@ private:
     float currentLeftSpeed;
     float currentRightSpeed;
 
-    int currentRightDistance;
-    int currentLeftDistance;
+    double currentRightDistance;
+    double currentLeftDistance;
 
     //Manual Mode variables (with joystick control)
     float requestedThrottle;
